REF: Split `Theme` into Colors.kt, Dimens.kt, Shapes.kt and Typography.kt
