+++ ui/src/main/java/cz/quanti/razym/rocketapp/ui/theme/Shapes.kt
@@ -1,0 +1,16 @@
+package cz.quanti.razym.rocketapp.ui.theme
+
+import androidx.compose.foundation.shape.CornerBasedShape
+import androidx.compose.material3.ShapeDefaults
+import androidx.compose.runtime.Immutable
+
+@Immutable
+class Shapes(
+    // Copy from MaterialTheme (commented out not used in app now).
+    // Shapes None and Full are omitted as None is a RectangleShape and Full is a CircleShape.
+//    val extraSmall: CornerBasedShape = ShapeDefaults.ExtraSmall,
+//    val small: CornerBasedShape = ShapeDefaults.Small,
+    val medium: CornerBasedShape = ShapeDefaults.Medium,
+//    val large: CornerBasedShape = ShapeDefaults.Large,
+//    val extraLarge: CornerBasedShape = ShapeDefaults.ExtraLarge,
+)
